axioma-api/requirements.txt
asgiref==3.11.1
boto3==1.42.97
botocore==1.42.97
dj-database-url==3.1.2
Django==6.0.4
django-cors-headers==4.9.0
django-environ==0.13.0
django-storages==1.14.6
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
jmespath==1.1.0
packaging==26.1
pillow==12.2.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dateutil==2.9.0.post0
s3transfer==0.16.1
six==1.17.0
sqlparse==0.5.5
tzdata==2026.1
urllib3==2.6.3
whitenoise==6.12.0
